separate class battlefield from the air_force_game file in a new file called air_force_battlefield
